write content and provider IDs on file projection
We pass the content and provider IDs to the libprojfs API function
projfs_create_proj_file(), which will store them as extended
attributes of the newly created file (pending revisions to the
libprojfs API).

For directory creation, however, we do not store any metadata
attributes, as only files may need to be subsequently hydrated with
data, in a second post-projection step which depends on the content ID
written during initial the projection step.