add pre-rename and post-delete Linux events
The GVFS.Virtualization.FileSystemCallbacks background task
processing code is responsible for adding and removing paths
to the GVFS modified-paths database, and in the case of rename
file operations, needs to remove the old path from the database.

In the Mac Darwin/BSD kauth design, a rename file operation is
preceded by a KAUTH_VNODE_DELETE permission check, which the
ProjFS.Mac kext reports as a PreDelete notification.  This in
turn ensures that the FileSystemCallbacks
ExecuteBackgroundOperation() method will remove the source path
from the modified-paths database.  However, the ProjFS.Linux
and libprojfs implementation did not include a parallel pre-delete
event prior to rename file operations, leading to functional test
failures on Linux.

Because there is no necessity to follow the BSD design, we
can introduce true pre-rename (PROJFS_MOVE_PERM) permission
requests in Linux, as well as post-delete events.  This brings
the libprojfs/ProjFS.Linux design somewhat closer to the Windows
projectedfslib/PrjFlt one.

In particular, this allows the GVFS LinuxFileSystemVirtualizer
to enforce restrictions on renaming the .git/index file, akin
to the ones enforced on Windows.

By reporting both the source and target (destination) paths in
the post-rename events, we permit the FileSystemCallbacks
task processing code to add and remove the paths to the database,
without having to do so in either the pre-rename event or a
BSD-like pre-delete event delivered prior to a rename.